Engine finds Checks and Pins
The engine now looks for attacks on the king and will not permit moves that place the king in check.

The isCheck, checks and pins logic is implimented in the 'check_for_pins_and_checks' function within the chess_engine.py file.

The check logic is also implimented in ALL of the piece specific move genration functions.  This varies from piece to piece and is unque for the king as it requires altering the king square and running that position throught the pins and checks function.

There have also been overhauls to the rook and bishop functions, reducing the complexity and length of the overall code to reduce issues in bug tracking.