bug fix: Pawn checks not registering and king unable to move out of check.
There was a type in the check pins and checks function.  One of the statements to set isCheck to True was written with == instead of =.  This was causing a bug where the pawn checks were not saving to the game state.

Second issue of the king unable to move out of check.  What was happening is that the king moves were being lumped in with other pieces when they were removed from the moves list if they did not block check.  Since the king does NOT need to block it's own check I added a check to only remove moves from the list if the moved piece was NOT a king piece.