bug fix:
In the event that a rook was captured before itself or the king had been moved, the option to castle was not set to false, meaning that the rook could be brought back to life by castling.

Added event to updateCastleRights that removes the ability to castle if the castle is captured by another piece.  Now when a move invloves the initial square of one of the rooks, the option to castle involving that rook is set to false.