Play as White
You can now flip the game board so that the white pieces are nearest to the player.  This can be toggeled with the playerIsWhite boolean.

No changes were made to the engine, all the changes were implimented within the gui application, so this should have no impact on the game mechanichs and logic.