Random Move AI
This includes the addition of a new file, ChessAI.py.  This file contains a python class with the temporary name of 'ComputerPlayer'.  The class has a single function which takes as an argument an object of type GameState and returns a move choice selected at random from the possible moves available.

This has been implimented in the ChessMain.py main function, is functional to play against as a single player game.