ai improvment:  Evaluate capture
The AI can now evaluate wether a series of exchanges will turn out in it's favour.  When a capture move is identified, a function 'evaluateCapture' looks to see the opposing pieces defening the square.  Then a calculation is done that measures what the 'best' possible outcome for the computer is.

The evaluation only measures material, position is NOT taken into account in the calculation, nor is there any thought given to defenses given up by moving the attacking pieces.