bug fix: pawn not moving on dxe5
There was a problem in the GUI for the entire duration of this project.  The issue was this:  When the player would click on a new square, after selecting one of his own pieces, the GUI would not recognize that the player had clicked on another piece, and a second click was required to "release" the original piece.

This was because of the confusing method which the row and column variables in the main account for the flipping of the board that occurs when looking from whites perspective.